Implemented ore displaying and ore mining

+++ Core/50rld/src/com/github/mimo31/w50rld/ItemStack.java
@@ -12,7 +12,7 @@ import java.awt.Rectangle;
 public class ItemStack {
 
 	// the number of items in this stack
-	private byte count = 0;
+	private int count = 0;
 	
 	// Item type stored
 	private Item item = null;
